pyproject.toml
[tool.poetry]
name = "dotfiles"
version = "0.1.0"
description = ""
authors = ["João Vitor Bonadiman <joao@bonadiman.dev>"]

[tool.poetry.dependencies]
python = "^3.9"
nanoid = "^2.0.0"
PyYAML = "^6.0"
requests = "^2.26.0"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
